Minor modification to the Spring autowwire tests
* Renamed class SpringContextTest inside both modules:
    - mats-test-junit
    - mats-test-jupiter
    to AbstractSpringTest
* Introduced inner classes acting as the Spring Configuration classes.
